add filter on extensions
It's now possible to add a filter on an extension.

atoum as a mechanism to ignore some tests if a php extension
is not loaded. For example :

```php
/**
 * @extensions blackfire
 */
class Example extends atoum
```

This change adds an extension variable to the filter rule.

You can now, for example launch only the tests that needs the blackfire
extension :

```
./vendor/bin/atoum -d tests --filter '"blackfire" in extensions'
```