improve README.md
In the previous versions all we can see without scrolling was atoum's logo
how to install the extensions. It will better to explain what the extension
does before installing.

Now we start with a little into, an example, and the list of filters, then
how to install.

After we have the examples.

The examples were not easly readable. There is now titles on the examples.

We have also added links to atoum's documentation.