Update Hoa dependency
Thanks @Hywan for the work he did on Hoa itself and for the initial PR (#29)

The first PR had mainly one issue: it forced us to drop support for PHP
5.4. We want to support as much PHP versions as possible.

The strategy here is the following:

* We add complex constraintis to composer requires matching libraries
  supporting 5.4 and those not supporting it. We do this by using a OR
  (`|`) in constraints,
* We added a `conflict` section to allow composer to correctly choose
  libraries it should install when the user asks for the last stable or
  the lowest possible deps,
* We added an autoloader to the extension (required to use the extension
  inside the dedicated docker image, atoum/atoum) which will take care
  of loading static files (for hoa/core, hoa/consistency and
  hoa/protocol) and will also provide an autoloader for classes in
  vendors and the extension itself.

Thanks to those three actions we can keep 5.4 support in the extension,
cover as much hoa and atoum versions as possible without conflict and let
composer do its job correctly. Then our autoloader will do the rest for
us.